third_party/blink/web_tests/media/media-fragments/TC0058-expected.txt


Title: TC0058
Fragment: 'T=3,7'
Comment: UA does not identify this as a media fragment. The entire media resource is played.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
END OF TEST
